servicedb/service/requirements.txt
click==8.1.3
colorama==0.4.6
databases==0.6.2
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
pyyaml==6.0
sniffio==1.3.0
sqlalchemy==1.4.41
starlette==0.21.0
typing-extensions==4.4.0
uvicorn[standard]==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
pika==1.3.1
asyncpg==0.27.0